requirements.txt
pyWinhook
pywin32
imageio
pynput
numpy
keyboard
mss
mouse
Pillow